move mock cache def to avoid circular imports, fix linter
